requirements.txt
langchain
langchain-google-genai
streamlit
sqlalchemy
python-dotenv
langchain-community